revise file name and move logic for start command
